Debugger: Fixed property_get/value data truncation and error-handling.
If the data exceeds max_data, truncate the data but return the total
size in the size attribute. This and some other cases previously caused
the response to indicate the property is empty.

If a memory allocation or string conversion fails, return an error 998
response instead of an incomplete (invalid) success response.

+++ source/Debugger.h
@@ -429,7 +429,7 @@ private:
 	int WritePropertyXml(ExprTokenType &aValue, const char *aName, CStringA &aNameBuf, int aPageSize, int aDepthRemaining, int aMaxEncodedSize);
 	int WritePropertyXml(Object::FieldType &aField, const char *aName, CStringA &aNameBuf, int aPageSize, int aDepthRemaining, int aMaxEncodedSize);
 
-	int WritePropertyData(LPCTSTR aData, int aDataSize, int aMaxEncodedSize);
+	int WritePropertyData(LPCTSTR aData, size_t aDataSize, int aMaxEncodedSize);
 	int WritePropertyData(Var &aVar, int aMaxEncodedSize);
 	int WritePropertyData(Object::FieldType &aField, int aMaxEncodedSize);
